requirements.txt
confluent_kafka
psycopg2-binary
requests
avro-python3
pandas
sqlmodel